requirements.txt
# Core ML libraries
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3

# Web frameworks
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2

# MLOps tools
mlflow==2.8.1
dvc==3.27.0
apache-airflow==2.7.3

# Data processing
requests==2.31.0
pydantic==2.5.0
Pillow==10.0.1

# Docker and deployment
docker==6.1.3

# Additional utilities
python-multipart==0.0.6
